Miscellaneous Fixes
Miscellaneous fixes to multiple files and refactored formatting to align with the rest of the project.

+++ src/sleep.c
@@ -2,10 +2,10 @@
 #include "pit.h"
 #include "sleep.h"
 
-void kernel_sleep_ms(uint64_t ms) {
-    uint64_t start = kernel_get_ticks_ms();
+void kernel_sleep_ms(uint32_t ms) {
+    uint32_t start = kernel_get_ticks_ms();
 
-    while ((kernel_get_ticks_ms() - start) < ms){
+    while ((kernel_get_ticks_ms() - start) < ms) {
         __asm__ volatile ("hlt");
     }
 }
