+++ src/sleep.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <stdint.h>
 
-void kernel_sleep_ms(uint64_t ms);
+void kernel_sleep_ms(uint32_t ms);
